requirements.txt
transformers
pandas
torch
spacy
gensim
sentence_transformers
nltk
bert-extractive-summarizer
sumy